+++ aksel.nav.no/website/pages/god-praksis/artikler/index.tsx
@@ -57,7 +57,7 @@ const Artikler = ({ articles }: ArtiklerProps) => {
         >
           <div className="relative grid overflow-x-clip pb-40">
             <AkselCubeStatic className="text-deepblue-300 opacity-5 " />
-            <div className="dynamic-wrapper centered-layout">
+            <div className="dynamic-wrapper mx-auto w-full px-4 sm:px-6">
               <Heading
                 level="1"
                 size="xlarge"