+++ aksel.nav.no/website/pages/produktbloggen/index.tsx
@@ -1,6 +1,13 @@
 import { Footer } from "@/layout";
-import { akselBloggPosts, SanityT, urlFor } from "@/lib";
+import { contributorsAll, destructureBlocks, urlFor } from "@/lib";
 import { getClient } from "@/sanity-client";
+import {
+  AkselBloggDocT,
+  AkselBloggFrontpageT,
+  NextPageT,
+  ResolveContributorsT,
+  ResolveSlugT,
+} from "@/types";
 import { Heading } from "@navikt/ds-react";
 import { Header } from "components/layout/header/Header";
 import BloggCard from "components/sanity-modules/cards/BloggCard";
@@ -12,7 +19,46 @@ import Head from "next/head";
 import { lazy } from "react";
 import NotFotfund from "../404";
 
-const Page = (props: PageProps): JSX.Element => {
+type PageProps = NextPageT<{
+  page: AkselBloggFrontpageT;
+  bloggposts: ResolveContributorsT<ResolveSlugT<AkselBloggDocT>>[];
+}>;
+
+export const query = `*[_type == "blogg_landingsside"][0]{
+  "page": {..., intro[]{...,${destructureBlocks}}},
+  "bloggposts": *[_type == "aksel_blogg"] | order(_createdAt desc){
+    seo,
+    heading,
+    ingress,
+    publishedAt,
+    _createdAt,
+    _id,
+    "slug": slug.current,
+    ${contributorsAll}
+  }
+}`;
+
+export const getStaticProps = async ({
+  preview = false,
+}: {
+  preview?: boolean;
+}): Promise<PageProps> => {
+  const { bloggposts, page } = await getClient().fetch(query);
+
+  return {
+    props: {
+      page,
+      bloggposts,
+      preview,
+      title: "Forside Blogg",
+      id: page?._id ?? "",
+    },
+    notFound: !bloggposts && !preview,
+    revalidate: 60,
+  };
+};
+
+const Page = (props: PageProps["props"]): JSX.Element => {
   if (!props.bloggposts) {
     return <NotFotfund />;
   }
@@ -57,7 +103,7 @@ const Page = (props: PageProps): JSX.Element => {
           className="relative min-h-[80vh] overflow-hidden focus:outline-none"
         >
           <AkselCubeStatic className="text-[#FFE78A] opacity-10" />
-          <div className="centered-layout mb-40 grid max-w-screen-2xl">
+          <div className="mx-auto mb-40 grid w-full max-w-screen-2xl px-4 sm:px-6">
             <LatestBloggposts
               bloggs={props?.bloggposts}
               title="Blogg"
@@ -93,7 +139,7 @@ const Wrapper = (props: any): JSX.Element => {
   if (props?.preview) {
     return (
       <PreviewSuspense fallback={<Page {...props} />}>
-        <WithPreview comp={Page} query={akselBloggPosts} props={props} />
+        <WithPreview comp={Page} query={query} props={props} />
       </PreviewSuspense>
     );
   }
@@ -102,42 +148,3 @@ const Wrapper = (props: any): JSX.Element => {
 };
 
 export default Wrapper;
-
-export type AkselBloggPage = Partial<
-  SanityT.Schema.aksel_blogg & {
-    slug: string;
-    contributors?: { title?: string }[];
-  }
->;
-
-interface PageProps {
-  bloggposts: AkselBloggPage[];
-  page: any;
-  preview: boolean;
-  title: string;
-}
-
-interface StaticProps {
-  props: PageProps;
-  notFound: boolean;
-  revalidate: number;
-}
-
-export const getStaticProps = async ({
-  preview = false,
-}: {
-  preview?: boolean;
-}): Promise<StaticProps | { notFound: true }> => {
-  const { bloggposts, page } = await getClient().fetch(akselBloggPosts);
-
-  return {
-    props: {
-      page,
-      bloggposts,
-      preview,
-      title: "Forside Blogg",
-    },
-    notFound: !bloggposts && !preview,
-    revalidate: 60,
-  };
-};