+++ aksel.nav.no/website/sanity/schema/documents/index.ts
@@ -20,4 +20,3 @@ export * from "./publiseringsflyt";
 export * from "./redirects";
 export * from "./skrivehjelp";
 export * from "./standalone-artikkel";
-export * from "./testDoc";