+++ aksel.nav.no/website/sanity/schema/index.ts
@@ -8,7 +8,6 @@ export const schema: SchemaPluginOptions = {
     document.Editors,
     document.Forside,
     document.Redirect,
-    document.TestDoc,
     document.Skrivehjelp,
     document.Publiseringsflyt,
     document.Feedback,
@@ -63,6 +62,7 @@ export const schema: SchemaPluginOptions = {
     object.RiktekstStandard,
     object.RiktekstPrinsipp,
     object.RiktekstStandalone,
+    object.RiktekstAccordion,
 
     /* Komponentsider */
     object.PropsSeksjon,