+++ @navikt/core/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/ds-react
 
+## 2.8.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/aksel-icons@2.8.4
+  - @navikt/ds-icons@2.8.4
+
 ## 2.8.3
 
 ### Patch Changes